got rid of scan and site select in favor of reuseable ListSelection component
