--- src/components/ScanSelect.vue
@@ -1,11 +1,0 @@
-<template>
-
-</template>
-
-<script setup>
-const props = defineProps({
-    scan: {}, //no requirements as it is set by component
-    dataset: {},
-    subject: {}
-})
-</script>
